refactor projects module
