+++ app/views/internal_api/v1/projects/create.json.jbuilder
@@ -1,3 +1,5 @@
 # frozen_string_literal: true
 
 json.partial! "project", locals: { project: }
+json.notice notice
+
