+++ app/views/internal_api/v1/projects/update.json.jbuilder
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 json.partial! "project", locals: { project: }
+json.notice notice
